misc: Use pre-commit for quality checks (#13361)
Summary:
This PR adds a [`.pre-commit-config.yaml`](https://pre-commit.com) with a number of hooks that mirror our previous `check.py` and some new ones to further improve code quality.

Hooks to replace check.py:
- cmake-format
- clang-format
- license header check
- ruff (python formatting and linting)
- clang-tidy (has to be called manually)

New Hooks:
- general quality checks
  - trailing white space removal
  - end of file fixer
  - large file check (is in the default config, we can probably drop it?)
  - check that files with shebangs are actually executable
  - and that executable scripts have shebangs
- Several checks to improve our GitHub Actions quality
  - zizmor, a security focused linter
  - action-validator, make sure actions are valid! (we have had issues with that before) also checks if paths globs are correct
  - yamlfmt/yamllint make sure we use a unified style in our yaml/workflows

I have added all fixes in this PR ~~but we could also merge them as a separate PR so it can be easier ignored when blaming.~~ Actually I think it is better to just handle everything in this PR to avoid causing a bunch of merge conflicts.

Compared to `check.py` this is also ***much*** faster. Running all hooks on all files in the repo takes under ten seconds!

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13361

Reviewed By: kKPulla

Differential Revision: D76274374

Pulled By: kgpai

fbshipit-source-id: ba23f75fd1ff6e05d0e2a41d791028ac0220da56